add percy to launchpad and frontend-shared
