+++ packages/frontend-shared/cypress/support/index.ts
@@ -1,6 +1,7 @@
-import { registerMountFn, addVueCommand } from './common'
+import { registerMountFn, addVueCommand, installCustomPercyCommand } from './common'
 import '../../src/styles/shared.scss'
 import 'virtual:windi.css'
 
+installCustomPercyCommand()
 registerMountFn()
 addVueCommand()
